tools/wave_comparison/pyproject.toml
[project]
    name = "wave-comparison"
    version = "0.0.1"
    description = "Compare wave files, which are similar"
    readme = "README.md"
    requires-python = ">=3.10"
    dependencies = [
        "librosa>=0.11.0",
        "numpy>=2.2.6",
        "scikit-learn>=1.7.0",
        "scipy>=1.15.3",
    ]

[project.optional-dependencies]
    cpu = [
        "faiss-cpu>=1.11.0",
    ]
    gpu = [
        "faiss-gpu>=1.7.2",
    ]
